+++ dev/openchpl_views.sql
@@ -1,4 +1,5 @@
 DROP VIEW IF EXISTS openchpl.questionable_activity_combined;
+DROP VIEW IF EXISTS openchpl.inactive_developers_and_products;
 DROP VIEW IF EXISTS openchpl.certified_product_search;
 DROP VIEW IF EXISTS openchpl.certified_product_details;
 DROP VIEW IF EXISTS openchpl.cqm_result_details;
@@ -136,6 +137,44 @@ CREATE OR REPLACE FUNCTION openchpl.get_active_listings_for_developer_during_per
     $$ LANGUAGE plpgsql
 stable;
 
+create or replace function openchpl.get_inactive_date_for_product(product_id_var bigint) returns
+    table (
+        inactive_date date
+        ) as $$
+    begin
+    return query
+        SELECT max(date(decert.listing_inactive_date))
+		FROM openchpl.certified_product cp
+		JOIN openchpl.product_version ver ON cp.product_version_id = ver.product_version_id
+		JOIN openchpl.product prod ON ver.product_id = prod.product_id
+		JOIN openchpl.vendor dev ON dev.vendor_id = prod.vendor_id
+		JOIN
+		  (SELECT max(certification_status_event.event_date) AS listing_inactive_date,
+				  certification_status_event.certified_product_id
+		   FROM openchpl.certification_status_event
+		   WHERE certification_status_event.certification_status_id NOT IN (1,6,7)
+		   AND certification_status_event.deleted = FALSE
+		   GROUP BY certification_status_event.certified_product_id) decert ON cp.certified_product_id = decert.certified_product_id
+		WHERE prod.product_id = product_id_var;
+end;
+$$ language plpgsql
+stable;
+
+create or replace function openchpl.get_acbs_for_product(product_id_var bigint) returns
+    table (
+        acbs_for_product text
+        ) as $$
+    begin
+    return query
+        SELECT string_agg(distinct acb.name, ';')
+		FROM openchpl.certified_product cp
+		JOIN openchpl.product_version ver ON cp.product_version_id = ver.product_version_id
+		JOIN openchpl.product prod ON ver.product_id = prod.product_id
+		JOIN openchpl.certification_body acb ON cp.certification_body_id = acb.certification_body_id
+		WHERE prod.product_id = product_id_var;
+end;
+$$ language plpgsql
+stable;
 
 CREATE VIEW openchpl.certification_result_details AS
 SELECT
@@ -1183,11 +1222,19 @@ FROM openchpl.certified_product cp
 		ON listing_search.certified_product_id = cp.certified_product_id
 		AND listing_search.certification_status_id IN (1,6,7);
 
+-- The 'requirement_type' view uses the magic number 70000.  This number was chosen because the 'requirement_group_type_id' is 7.
+-- If we need to union more data into this view, we should use a similar pattern.  This gives use a range of 10000 
+-- certification_criterion_id before we have an issue with this view.
 CREATE OR REPLACE VIEW openchpl.requirement_type
 AS
+SELECT certification_criterion_id + 70000 as id, title, number, start_day, end_day, certification_edition_id, 7 as requirement_group_type_id
+FROM openchpl.certification_criterion
+WHERE (certification_edition_id IS NULL
+	OR certification_edition_id != (SELECT certification_edition_id FROM openchpl.certification_edition where year = '2011'))
+UNION
 SELECT certification_criterion_id as id, title, number, start_day, end_day, certification_edition_id, 1 as requirement_group_type_id
 FROM openchpl.certification_criterion
-WHERE (certification_edition_id IS NULL 
+WHERE (certification_edition_id IS NULL
 	OR certification_edition_id != (SELECT certification_edition_id FROM openchpl.certification_edition where year = '2011'))
 UNION
 SELECT id, name, null, null, null, null, requirement_group_type_id
@@ -1427,6 +1474,19 @@ LEFT JOIN (SELECT string_agg(certification_body_id::text||':'||name, '|') as acb
 		GROUP BY vendor_id) dev_acb_map2
 	    ON dev_acb_map2.vendor_id = dev.vendor_id		
 WHERE dev.deleted = false;
+
+CREATE OR REPLACE VIEW openchpl.inactive_developers_and_products
+AS 
+SELECT vendor_id, vendor_name, vendor_website, product_id, product_name, 
+	openchpl.get_acbs_for_product(product_id) as certification_bodies, 
+  openchpl.get_inactive_date_for_product(product_id) as inactive_date
+FROM openchpl.certified_product_details cpd
+JOIN openchpl.developer_search ds
+	ON ds.developer_id = cpd.vendor_id
+  AND ds.current_active_listing_count = 0
+WHERE cpd.deleted = false
+GROUP BY vendor_id, vendor_name, vendor_website, product_id, product_name
+ORDER BY vendor_id, product_id;
 
 CREATE VIEW openchpl.subscription_search_result AS
  SELECT subscriber.id as subscriber_id,